- fixed wrong comments in phystut - fixed compilation warning in dynsysdebug - cleaned a bit creation of bodymesh
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@33429 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tutorial/phystut/dynsysdebug.cpp
@@ -225,8 +225,8 @@ void ColoredTexture::Animate (csTicks current_time)
 {
   g3d->SetRenderTarget (GetTextureWrapper ()->GetTextureHandle ());
   if (!g3d->BeginDraw(CSDRAW_2DGRAPHICS)) return;
-  g3d->GetDriver2D()->DrawPixel (0, 0, g3d->GetDriver2D()->FindRGB (
-            color.red * 255.0, color.green * 255.0, color.blue * 255.0));
+  g3d->GetDriver2D()->DrawPixel (0, 0, g3d->GetDriver2D()->FindRGB
+         ((int) (color.red * 255.0), (int) (color.green * 255.0), (int) (color.blue * 255.0)));
   g3d->FinishDraw ();
 }
